Consolidate Pilot tests to reduce test time
